refactor: code conduct
