+++ src/components/settings-account-style.tsx
@@ -1,4 +1,4 @@
-import React, {useEffect} from 'react';
+import React, {useCallback, useEffect, useRef, useState} from 'react';
 
 import {useWindowDimensions} from 'react-native';
 import Animated, {
@@ -20,8 +20,11 @@ import {
 } from '@app/components/ui';
 import {createTheme} from '@app/helpers';
 import {I18N} from '@app/i18n';
+import {Wallet} from '@app/models/wallet';
+import {HapticEffects, vibrate} from '@app/services/haptic';
 import {WalletCardPattern, WalletCardStyle} from '@app/types';
-
+import {generateFlatColors, generateGradientColors} from '@app/utils';
+import {CARD_CIRCLE_TOTAL, CARD_RHOMBUS_TOTAL} from '@app/variables';
 const cardStyleVariants = [
   {
     value: WalletCardStyle.flat,
@@ -50,32 +53,35 @@ const patternVariants = [
 
 type SettingsAccountStyleProps = {
   onPressApply: () => void;
-  onPressGenerate: () => void;
-  onChangePattern: (value: WalletCardPattern) => void;
-  onChangeType: (value: WalletCardStyle) => void;
   patternStyle: WalletCardPattern;
   colors: string[];
   cardStyle: WalletCardStyle;
   pattern: string;
-  loading: boolean;
-  isStyleChanged: boolean;
+  setPattern: (value: string) => void;
+  setColors: (value: string[]) => void;
+  setPatternStyle: (value: WalletCardPattern) => void;
+  setCardStyle: (value: WalletCardStyle) => void;
+  wallet: Wallet;
 };
 
 export const SettingsAccountStyle = ({
   onPressApply,
-  onPressGenerate,
-  onChangePattern,
-  onChangeType,
   patternStyle,
   cardStyle,
   colors,
   pattern,
-  loading,
-  isStyleChanged,
+  setPattern,
+  setColors,
+  setPatternStyle,
+  setCardStyle,
+  wallet,
 }: SettingsAccountStyleProps) => {
   const opacity = useSharedValue(1);
   const heightCard = 192;
+  const timerRef: {current: NodeJS.Timeout | null} = useRef(null);
   const cardWidth = useWindowDimensions().width - 72;
+  const [loading, setLoading] = useState<boolean>(false);
+  const [isStyleChanged, setIsStyleChanged] = useState(false);
 
   const animatedStyles = useAnimatedStyle(() => {
     return {
@@ -87,6 +93,93 @@ export const SettingsAccountStyle = ({
     opacity.value = withTiming(loading ? 0.5 : 1, {duration: 500});
   }, [loading, opacity]);
 
+  useEffect(() => {
+    return () => {
+      if (timerRef.current) {
+        clearTimeout(timerRef.current);
+      }
+    };
+  }, []);
+
+  useEffect(() => {
+    const {
+      cardStyle: WCardStyle,
+      colorFrom,
+      colorTo,
+      colorPattern,
+      pattern: WPattern,
+    } = wallet;
+    if (
+      WCardStyle !== cardStyle ||
+      colors[0] !== colorFrom ||
+      colors[1] !== colorTo ||
+      colors[2] !== colorPattern ||
+      WPattern !== pattern
+    ) {
+      !isStyleChanged && setIsStyleChanged(true);
+    } else {
+      isStyleChanged && setIsStyleChanged(false);
+    }
+  }, [isStyleChanged, colors, cardStyle, wallet, pattern]);
+
+  const onPressGenerate = useCallback(() => {
+    vibrate(HapticEffects.impactLight);
+    setLoading(true);
+    const newColors =
+      cardStyle === WalletCardStyle.flat
+        ? generateFlatColors()
+        : generateGradientColors();
+
+    const newPattern = `${patternStyle}-${Math.floor(
+      Math.random() *
+        (patternStyle === WalletCardPattern.circle
+          ? CARD_CIRCLE_TOTAL
+          : CARD_RHOMBUS_TOTAL),
+    )}`;
+    timerRef.current = setTimeout(() => {
+      setPattern(newPattern);
+      setColors(newColors);
+      setLoading(false);
+    }, 500);
+  }, [cardStyle, patternStyle, setColors, setPattern]);
+
+  const onChangePattern = useCallback(
+    (value: WalletCardPattern) => {
+      if (pattern !== value) {
+        setPatternStyle(value);
+
+        const newPattern = `${value}-${Math.floor(
+          Math.random() *
+            (value === WalletCardPattern.circle
+              ? CARD_CIRCLE_TOTAL
+              : CARD_RHOMBUS_TOTAL),
+        )}`;
+
+        setPattern(newPattern);
+        if (patternStyle !== value) {
+          opacity.value = withTiming(0.5, {duration: 250}, finished => {
+            if (finished) {
+              opacity.value = withTiming(1, {duration: 250});
+            }
+          });
+        }
+      }
+    },
+    [pattern, opacity, patternStyle, setPattern, setPatternStyle],
+  );
+
+  const onChangeType = useCallback(
+    (value: WalletCardStyle) => {
+      setCardStyle(value);
+      const newColors =
+        value === WalletCardStyle.flat
+          ? generateFlatColors()
+          : generateGradientColors();
+
+      setColors(newColors);
+    },
+    [setCardStyle, setColors],
+  );
   return (
     <PopupContainer style={styles.container}>
       <Animated.View style={animatedStyles}>
